Cambiar js con modulo para cumplir con el ejercicio 22 por Gabriel Calisaya
