+++ assets/pages/ejercicio22.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Ejercicio 22 - Programacion Visual</title>
-    <link rel="stylesheet" href="/PV_Tp2_Grupo24/estilos/estiloEjercicio22.css">
+    <link rel="stylesheet" href="/estilos/estiloEjercicio22.css">
 </head>
 
 <body>
@@ -19,7 +19,7 @@
         <h2>Tu texto es:</h2>
         <div id="salidaTexto"></div>
     </div>
-    <script src="/PV_Tp2_Grupo24/src/ejercicio22.js"></script>
+    <script type="module" src="/src/ejercicio22.js"></script>
 </body>
 
 </html>
