+++ estilos/estiloEjercicio22.css
@@ -56,5 +56,4 @@ h3 {
     background-color: rgb(243, 59, 59);
     color: rgb(255, 255, 255);
     font-weight: bold;
-
 }
