+++ src/ejercicio22.js
@@ -1,12 +1,3 @@
-const ingresoTexto = document.getElementById('ingresoTexto');
-const salidaTexto = document.getElementById('salidaTexto');
+import { inicializarEjercicio22 } from '/src/modulo_ejercicio22.js';
 
-ingresoTexto.addEventListener('input', function () {
-    salidaTexto.textContent = ingresoTexto.value;
-
-    if (ingresoTexto.value.length > 20) {
-        salidaTexto.classList.add('cambiar-fondo');
-    } else {
-        salidaTexto.classList.remove('cambiar-fondo');
-    }
-});
+inicializarEjercicio22();
